s3: make profile and locking optional

+++ nixops/storage/s3.py
@@ -5,7 +5,7 @@ from botocore.exceptions import ClientError
 import sys
 import os
 import typing
-from typing import Dict
+from typing import Dict, Optional
 
 if typing.TYPE_CHECKING:
     import nixops.statefile
@@ -17,13 +17,13 @@ class S3Backend:
         raise NotImplementedError
 
     def __init__(self, args: StorageArgValues) -> None:
-        self.bucket = args["bucket"]
-        self.key = args["key"]
-        self.region = args["region"]
-        self.profile = args.get("profile")
-        self.dynamodb_table = args["dynamodb_table"]
-        self.s3_endpoint = args.get("s3_endpoint")
-        self.kms_keyid = args.get("kms_keyid")
+        self.bucket: str = args["bucket"]
+        self.key: str = args["key"]
+        self.region: str = args["region"]
+        self.profile: Optional[str] = args.get("profile")
+        self.dynamodb_table: Optional[str] = args.get("dynamodb_table")
+        self.s3_endpoint: Optional[str] = args.get("s3_endpoint")
+        self.kms_keyid: Optional[str] = args.get("kms_keyid")
         self.aws = boto3.Session(region_name=self.region, profile_name=self.profile)
 
     # fetchToFile: acquire a lock and download the state file to
@@ -69,14 +69,16 @@ class S3Backend:
             return {}
 
     def lock(self, path) -> None:
-        r = self.aws.client("dynamodb").put_item(
-            TableName=self.dynamodb_table,
-            Item={"LockID": {"S": f"{self.bucket}/{self.key}"},},
-            ConditionExpression="attribute_not_exists(LockID)",
-        )
+        if self.dynamodb_table is not None:
+            self.aws.client("dynamodb").put_item(
+                TableName=self.dynamodb_table,
+                Item={"LockID": {"S": f"{self.bucket}/{self.key}"},},
+                ConditionExpression="attribute_not_exists(LockID)",
+            )
 
     def unlock(self, path: str) -> None:
-        self.aws.client("dynamodb").delete_item(
-            TableName=self.dynamodb_table,
-            Key={"LockID": {"S": f"{self.bucket}/{self.key}"},},
-        )
+        if self.dynamodb_table is not None:
+            self.aws.client("dynamodb").delete_item(
+                TableName=self.dynamodb_table,
+                Key={"LockID": {"S": f"{self.bucket}/{self.key}"},},
+            )
